Melhorado responsividade na pagina comunidade.php e separado em arquivos comunidade.css e os scripts dos graficos
